feat(scheduler): add skeleton for the scheduler
This adds a new binary crate, `orka-scheduler`, which is the application that takes the role of the Scheduler in Orka's architecture.

What's implemented?
===================

Start up
--------

When the scheduler starts up, it will first parse the command line arguments (described thereafter) and set the verbosity level of the logger.

Afterwards, all the necessary directories are created on the filesystem.

TLS secrets management
----------------------

Then, if TLS has not been disabled via the configuration, it is prepared for being used with gRPC:
- First, the private key and certificate are read from `$DATA_DIR/tls/scheduler.{key,pem}`
- If the files could not be read (files not present or permission denied), and TLS secrets generation has not been disabled via the configuration, then a private key and a self-signed certificate are generated
- Both files are then written to the disk for persistence

This step can fail if the private key or certificate could not:
- be read, and TLS secrets generation is disabled
- be generated
- be written to the disk (only after being generated)

The self-signed certificate is generated using the following parameters:
- `CN = Orka Scheduler TLS certificate; O = Orka; OU = Scheduler`
- Not before: server date minus one (1) day
- Not after: server date plus ten (10) years

gRPC server
-----------

Finally, the gRPC server manager is created using the bind address and port retrieved from the configuration, and the TLS manager previously configured if it was not disabled.

The gRPC server itself will immediately be configured with:
- The TLS private key and certificate, if enabled
- Servers for the `scheduler.agent.LifecycleService`, `scheduler.agent.StatusUpdateService` and `scheduler.controller.SchedulingService` services
- The socket address to bind the server to

The gRPC server is then started, serving requests until the process is stopped.

For now, the methods in the gRPC services are only stubs that will be implemented later.

Scheduler configuration
=======================

The scheduler can be configured through command line arguments (or environment variables, which have less priority):
- `--data-dir`: directory where the scheduler will store its data (default: `/var/lib/orka/scheduler/` - the parent directory has been discussed and is expected to be used by all Orka services)
- `--no-tls`: disable TLS for the gRPC server
- `--no-tls-secret-generation`: disable the generation of the certificate and private key for TLS if none has been found in the data directory
- `--grpc-bind-address`: the address to bind the gRPC server to
- `--grpc-bind-port`: the port to bind the gRPC server to
- `-v`, `--verbose`, `-q`, `--quiet`: increase/decrease the verbosity level of the logger

Signed-off-by: Kuruyia <8174691+kuruyia@users.noreply.github.com>