feat(scheduler): implement node agent management
This adds the implementation for the different services referenced in the `agent.proto` to the `orka-scheduler` crate.

# Implementations
## Managers
The Scheduler now features a Node Agent Manager, responsible for the management and storage of the registered Agents. Custom errors have been implemented for better readability.

Each Node Agent contains the following members:
* a heartbeat, in order for the Scheduler to know which Agents are alive and responding;
* CPU metrics, which for now contain only the CPU load percentage;
* Memory metrics, which include total memory, and available memory, both in bytes.

## Node agent lifecycle
### Joining the cluster
The Node Agent can connect to the Scheduler, then get registered against the Scheduler by giving its ID. If an Agent is already registered with this ID, the response status will be `ALREADY_EXISTS` (status code `6`).

The Scheduler is not responsible for Agent ID generation. However, we do advise:
* to generate the ID for each Agent process, rather than for each Node the Agent is installed onto;
* to generate an ID using an already-established format that will lead to a unique ID, such as UUIDv4.

**Note**: Agents are trusted by default because TLS is used for authenticating and securing the connection. If a user disables TLS, it is their responsibility to ensure that Agents can be trusted.

### Leaving the cluster
Upon a graceful exit, the Agent must send a disconnection notice to the Scheduler. This notice is always answered with an empty response and an OK status code, and the Agent can ignore this response.

## Node agent status updates
Right after its registration, the Agent must call the node status update endpoint. This will open a stream that the Agent must try to keep open for its entire lifetime.

Periodically, the Agent sends a status update, containing the metrics of the Node it is installed on. Those are listed in the [Managers section](#managers). This also allows the Scheduler to ensure the Agent is still alive and well.

Should the endpoint be called by an unregistered or unknown Agent, the stream will be closed with a response status of `NOT_FOUND` (status code `5`).

Signed-off-by: BioTheWolff <47079795+BioTheWolff@users.noreply.github.com>