feat(scheduler): add workload management and scheduler functionalities
This adds the scheduling logic for the Orka's Scheduler.

# Implementation
## Scheduling requests
The Scheduler can handle incoming scheduling requests from the Controller. The Scheduler then decides of the fitting Agent, and contacts to assign it the workload. Two one-way streams are then established: one from the Node Agent to the Scheduler, and the other from the Scheduler to the Controller. Every workload status update sent by the Agent is then re-mapped and forwarded to the Controller.

The stream to the Controller is closed after the Agent's stream was closed. As the Scheduler never closes said streams itself, an unexpected end of the Scheduler-Controller stream must cause an error or a warning in the Controller. It is the Controller's responsibility to re-open the streams, or - upon re-gaining contact with the Scheduler - to send the appropriate signal to terminate the Workload.

## Workload signals
The Scheduler implements two methods, `stop` and `destroy`, that allow the Controller to request the respective signals to be sent to the workload. Said signal is thus forwarded by the Scheduler to the corresponding Node Agent, according to the association `instance -> agent` that the Scheduler possesses. An error is thrown if the instance ID is unknown, unbound (no corresponding Agent that hosts the workload) or if the Agent sends an error. In case of an error, the Controller must assume the signal has not gone through and the workload is still running.

---------

Co-authored-by: BioTheWolff <47079795+BioTheWolff@users.noreply.github.com>
Co-authored-by: Adrien Raimbault <adrienraimbault@hotmail.com>
Co-authored-by: Kuruyia <8174691+kuruyia@users.noreply.github.com>
Signed-off-by: BioTheWolff <47079795+BioTheWolff@users.noreply.github.com>