feat(node-agent): implement container creation
Add a wrapper for containerd-client lib (and ctr cli for creation) to create a container and run a workload.

This will however mean more resource usage, overhead, latency, potential socket exhaustion, etc.. This is a tradeoff we are willing to make for now, since these issues are not likely to be a problem in the near future.

The ctr cli was used to create the container because the containerd-client we did not manage to create a container with the containerd-client lib. It runs the command ctr run -d <image> <container-id>.

**Note**: It seems Kuruyia has managed to prepare a rootfs and run a container, you can find it [here](https://github.com/Martin-Moreira-de-jesus/orka/blob/tmp/containerd-create-workload/node-agent/src/workload_manager/container/client.rs).

Also note that for some reason if you dont build the node-agent binary from the folder node-agent, the containerd socket cannot be found for some reason.

Signed-off-by: Martin Moreira de Jesus <martin.moreiradj@icloud.com>
Signed-off-by: Noé Tarbouriech <noe.tarbouriech@etu.umontpellier.fr>
Co-Authored-By: Noé Tarbouriech <noe.tarbouriech@etu.umontpellier.fr>