feat(node-agent): add status, cleanup and kill
The `containerd-client` is used to kill the container. We did however have an issue concerning an ambiguous type definition. Indeed, the Workload Signal enum is ambiguous. It is not clear what signal STOP represents. For now we made the assumption that stop is SIGINT (2). Both Stop (SIGINT) and Kill (SIGKILL) signals are sent to the workload, then the workload is cleaned up.

Signed-off-by: Martin Moreira de Jesus <martin.moreiradj@icloud.com>
Signed-off-by: Noé Tarbouriech <noe.tarbouriech@etu.umontpellier.fr>
Co-Authored-By: Noé Tarbouriech <noe.tarbouriech@etu.umontpellier.fr>